refactor: move config to separate file
